+++ weblate/trans/tests/test_machine.py
@@ -227,13 +227,19 @@ TERMINOLOGY_TRANSLATE = '''
 
 class MachineTranslationTest(TestCase):
     """Testing of machine translation core."""
+    def get_machine(self, cls):
+        machine = cls()
+        machine.delete_cache()
+        machine.cache_translations = False
+        return machine
+
     def test_support(self):
-        machine_translation = DummyTranslation()
+        machine_translation = self.get_machine(DummyTranslation)
         self.assertTrue(machine_translation.is_supported('en', 'cs'))
         self.assertFalse(machine_translation.is_supported('en', 'de'))
 
     def test_translate(self):
-        machine_translation = DummyTranslation()
+        machine_translation = self.get_machine(DummyTranslation)
         self.assertEqual(
             machine_translation.translate('cs', 'Hello', MockUnit(), None),
             []
@@ -248,7 +254,7 @@ class MachineTranslationTest(TestCase):
         )
 
     def test_translate_fallback(self):
-        machine_translation = DummyTranslation()
+        machine_translation = self.get_machine(DummyTranslation)
         self.assertEqual(
             len(
                 machine_translation.translate(
@@ -259,7 +265,7 @@ class MachineTranslationTest(TestCase):
         )
 
     def test_translate_fallback_missing(self):
-        machine_translation = DummyTranslation()
+        machine_translation = self.get_machine(DummyTranslation)
         self.assertEqual(
             machine_translation.translate(
                 'de_CZ', 'Hello, world!', MockUnit(), None
@@ -275,8 +281,7 @@ class MachineTranslationTest(TestCase):
 
     @httpretty.activate
     def test_glosbe(self):
-        machine = GlosbeTranslation()
-        machine.delete_cache()
+        machine = self.get_machine(GlosbeTranslation)
         httpretty.register_uri(
             httpretty.GET,
             'https://glosbe.com/gapi/translate',
@@ -287,8 +292,7 @@ class MachineTranslationTest(TestCase):
 
     @httpretty.activate
     def test_glosbe_ratelimit(self):
-        machine = GlosbeTranslation()
-        machine.delete_cache()
+        machine = self.get_machine(GlosbeTranslation)
         httpretty.register_uri(
             httpretty.GET,
             'https://glosbe.com/gapi/translate',
@@ -301,8 +305,7 @@ class MachineTranslationTest(TestCase):
 
     @httpretty.activate
     def test_glosbe_ratelimit_set(self):
-        machine = GlosbeTranslation()
-        machine.delete_cache()
+        machine = self.get_machine(GlosbeTranslation)
         machine.set_rate_limit()
         httpretty.register_uri(
             httpretty.GET,
@@ -314,8 +317,7 @@ class MachineTranslationTest(TestCase):
     @override_settings(MT_MYMEMORY_EMAIL='test@weblate.org')
     @httpretty.activate
     def test_mymemory(self):
-        machine = MyMemoryTranslation()
-        machine.delete_cache()
+        machine = self.get_machine(MyMemoryTranslation)
         httpretty.register_uri(
             httpretty.GET,
             'https://mymemory.translated.net/api/get',
@@ -326,8 +328,7 @@ class MachineTranslationTest(TestCase):
 
     @httpretty.activate
     def test_apertium(self):
-        machine = ApertiumTranslation()
-        machine.delete_cache()
+        machine = self.get_machine(ApertiumTranslation)
         httpretty.register_uri(
             httpretty.GET,
             'http://api.apertium.org/json/listPairs',
@@ -346,8 +347,7 @@ class MachineTranslationTest(TestCase):
     @override_settings(MT_APERTIUM_APY='http://apertium.example.com/')
     @httpretty.activate
     def test_apertium_apy(self):
-        machine = ApertiumAPYTranslation()
-        machine.delete_cache()
+        machine = self.get_machine(ApertiumAPYTranslation)
         httpretty.register_uri(
             httpretty.GET,
             'http://apertium.example.com/listPairs',
@@ -366,8 +366,7 @@ class MachineTranslationTest(TestCase):
     @override_settings(MT_MICROSOFT_ID='ID', MT_MICROSOFT_SECRET='SECRET')
     @httpretty.activate
     def test_microsoft(self):
-        machine = MicrosoftTranslation()
-        machine.delete_cache()
+        machine = self.get_machine(MicrosoftTranslation)
         httpretty.register_uri(
             httpretty.POST,
             'https://datamarket.accesscontrol.windows.net/v2/OAuth2-13',
@@ -391,8 +390,7 @@ class MachineTranslationTest(TestCase):
     @override_settings(MT_MICROSOFT_COGNITIVE_KEY='KEY')
     @httpretty.activate
     def test_microsoft_cognitive(self):
-        machine = MicrosoftCognitiveTranslation()
-        machine.delete_cache()
+        machine = self.get_machine(MicrosoftCognitiveTranslation)
         httpretty.register_uri(
             httpretty.POST,
             'https://api.cognitive.microsoft.com/sts/v1.0/issueToken'
@@ -421,8 +419,7 @@ class MachineTranslationTest(TestCase):
                 return (200, headers, TERMINOLOGY_LANGUAGES)
             return (200, headers, TERMINOLOGY_TRANSLATE)
 
-        machine = MicrosoftTerminologyService()
-        machine.delete_cache()
+        machine = self.get_machine(MicrosoftTerminologyService)
         httpretty.register_uri(
             httpretty.POST,
             'http://api.terminology.microsoft.com/Terminology.svc',
@@ -434,8 +431,7 @@ class MachineTranslationTest(TestCase):
 
     @httpretty.activate
     def test_microsoft_terminology_error(self):
-        machine = MicrosoftTerminologyService()
-        machine.delete_cache()
+        machine = self.get_machine(MicrosoftTerminologyService)
         httpretty.register_uri(
             httpretty.POST,
             'http://api.terminology.microsoft.com/Terminology.svc',
@@ -449,8 +445,7 @@ class MachineTranslationTest(TestCase):
     @override_settings(MT_GOOGLE_KEY='KEY')
     @httpretty.activate
     def test_google(self):
-        machine = GoogleTranslation()
-        machine.delete_cache()
+        machine = self.get_machine(GoogleTranslation)
         httpretty.register_uri(
             httpretty.GET,
             GOOGLE_API_ROOT + 'languages',
@@ -479,8 +474,7 @@ class MachineTranslationTest(TestCase):
     @httpretty.activate
     def test_google_invalid(self):
         """Test handling of server failure."""
-        machine = GoogleTranslation()
-        machine.delete_cache()
+        machine = self.get_machine(GoogleTranslation)
         httpretty.register_uri(
             httpretty.GET,
             'https://www.googleapis.com/language/translate/v2/languages',
@@ -498,8 +492,7 @@ class MachineTranslationTest(TestCase):
 
     @httpretty.activate
     def test_amagama_nolang(self):
-        machine = AmagamaTranslation()
-        machine.delete_cache()
+        machine = self.get_machine(AmagamaTranslation)
         httpretty.register_uri(
             httpretty.GET,
             'https://amagama-live.translatehouse.org/api/v1/languages/',
@@ -521,8 +514,7 @@ class MachineTranslationTest(TestCase):
 
     @httpretty.activate
     def test_amagama(self):
-        machine = AmagamaTranslation()
-        machine.delete_cache()
+        machine = self.get_machine(AmagamaTranslation)
         httpretty.register_uri(
             httpretty.GET,
             'https://amagama-live.translatehouse.org/api/v1/languages/',
@@ -544,8 +536,7 @@ class MachineTranslationTest(TestCase):
     @override_settings(MT_YANDEX_KEY='KEY')
     @httpretty.activate
     def test_yandex(self):
-        machine = YandexTranslation()
-        machine.delete_cache()
+        machine = self.get_machine(YandexTranslation)
         httpretty.register_uri(
             httpretty.GET,
             'https://translate.yandex.net/api/v1.5/tr.json/getLangs',
@@ -562,8 +553,7 @@ class MachineTranslationTest(TestCase):
     @override_settings(MT_YANDEX_KEY='KEY')
     @httpretty.activate
     def test_yandex_error(self):
-        machine = YandexTranslation()
-        machine.delete_cache()
+        machine = self.get_machine(YandexTranslation)
         httpretty.register_uri(
             httpretty.GET,
             'https://translate.yandex.net/api/v1.5/tr.json/getLangs',
@@ -580,8 +570,7 @@ class MachineTranslationTest(TestCase):
     @override_settings(MT_SAP_BASE_URL='http://sth.example.com/')
     @httpretty.activate
     def test_saptranslationhub(self):
-        machine = SAPTranslationHub()
-        machine.delete_cache()
+        machine = self.get_machine(SAPTranslationHub)
         httpretty.register_uri(
             httpretty.GET,
             'http://sth.example.com/languages',
@@ -616,8 +605,7 @@ class MachineTranslationTest(TestCase):
     @override_settings(MT_SAP_BASE_URL='http://sth.example.com/')
     @httpretty.activate
     def test_saptranslationhub_invalid(self):
-        machine = SAPTranslationHub()
-        machine.delete_cache()
+        machine = self.get_machine(SAPTranslationHub)
         httpretty.register_uri(
             httpretty.GET,
             'http://sth.example.com/languages',
@@ -636,14 +624,32 @@ class MachineTranslationTest(TestCase):
     @override_settings(MT_DEEPL_KEY='KEY')
     @httpretty.activate
     def test_deepl(self):
-        machine = DeepLTranslation()
-        machine.delete_cache()
+        machine = self.get_machine(DeepLTranslation)
         httpretty.register_uri(
             httpretty.POST,
             'https://api.deepl.com/v1/translate',
             body=b'{ "translations": [ { "detected_source_language": "EN", "text": "Hallo" } ] }'
         )
         self.assert_translate(machine, lang='de', word='Hello')
+
+    @override_settings(MT_DEEPL_KEY='KEY')
+    @httpretty.activate
+    def test_cache(self):
+        machine = self.get_machine(DeepLTranslation)
+        machine.cache_translations = True
+        httpretty.register_uri(
+            httpretty.POST,
+            'https://api.deepl.com/v1/translate',
+            body=b'{ "translations": [ { "detected_source_language": "EN", "text": "Hallo" } ] }'
+        )
+        # Fetch from service
+        self.assert_translate(machine, lang='de', word='Hello')
+        self.assertTrue(httpretty.has_request())
+        httpretty.reset()
+        # Fetch from cache
+        self.assert_translate(machine, lang='de', word='Hello')
+        self.assertFalse(httpretty.has_request())
+
 
 
 class WeblateTranslationTest(FixtureTestCase):