+++ docs/changes.rst
@@ -20,6 +20,7 @@ Released on ? 2018.
 * Added built in translation memory.
 * Added componentlists overview to dashboard and per component list overview pages.
 * Added support for DeepL machine translation service.
+* Machine translation results are now cached inside Weblate.
 
 weblate 2.19.1
 --------------